Fixes countdown in story; bug was worse than I thought.
Version as submitted either requires perfect play after entering the
comm car or may be impossible to win. =((( Stopping the computer
didn't stop the countdown. Bug in the story. So you'd have 10 actions
after entering the comm car to stop the computer and leave the train.

This properly disables the countdown after stopping the computer.